+++ lib/configuration.js
@@ -1,30 +1,13 @@
 "use strict";
 
-const targets = {
-  csharp: {
-    lang: "C#",
-    extension: "cshtml",
-    attr: {
-      model: "@model",
-      using: "@using"
-    }
-  },
-  vb: {
-    lang: "VB",
-    extension: "vbhtml",
-    attr: {
-      model: "@ModelType",
-      using: "@Imports"
-    }
-  }
-};
+const targets = require("./targets");
 
 let _rules;
 
 class Configuration {
   constructor(opts, rules) {
     const options = typeof opts === "object" ? opts : {};
-    this.target = this._getTarget(options);
+    this.target = targets.resolve(options.target || "");
     this.rules = rules;
   }
